Added more error cases for edit_schedule.

+++ main.py
@@ -84,8 +84,10 @@ def create_schedule(schedule: Schedule):
 @app.post("/edit-schedule")
 def edit_schedule(schedule: ScheduleEdit):
     """Delete the old schedule and create a new one. If the new schedule information already exists, return an error."""
+    if schedule_collection.find_one({"day_name": schedule.old_day_name, "time": schedule.old_time}) is None:
+        return {"status": "The schedule targeted for change does not exist!"}
     if schedule_collection.find_one({"day_name": schedule.day_name, "time": schedule.time}) is not None:
-        return {"status": "Schedule already exists!"}
+        return {"status": "The provided new schedule combination already exists!"}
     schedule_collection.delete_one({"day_name": schedule.old_day_name, "time": schedule.old_time})
     schedule_collection.insert_one(schedule.dict())
     return {"status": "Schedule edited!",
